Move domain_languages.common to semparse.common
