--- allennlp/semparse/domain_languages/common/__init__.py
@@ -1,1 +1,0 @@
-from allennlp.semparse.domain_languages.common.date import Date
